+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/AssistantCreateParamsTest.kt
@@ -61,7 +61,8 @@ internal class AssistantCreateParamsTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(
@@ -276,7 +277,8 @@ internal class AssistantCreateParamsTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(
@@ -495,7 +497,8 @@ internal class AssistantCreateParamsTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(