+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/AssistantUpdateParamsTest.kt
@@ -63,7 +63,8 @@ internal class AssistantUpdateParamsTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(
@@ -289,7 +290,8 @@ internal class AssistantUpdateParamsTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(
@@ -509,7 +511,8 @@ internal class AssistantUpdateParamsTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(