+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/InferenceEmbeddingTest.kt
@@ -75,7 +75,8 @@ internal class InferenceEmbeddingTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(
@@ -303,7 +304,8 @@ internal class InferenceEmbeddingTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(
@@ -556,7 +558,8 @@ internal class InferenceEmbeddingTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(