+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/TelephonySettingsTest.kt
@@ -30,7 +30,8 @@ internal class TelephonySettingsTest {
                 )
                 .supportsUnauthenticatedWebCalls(true)
                 .timeLimitSecs(30L)
-                .userIdleTimeoutSecs(30L)
+                .userIdleReplySecs(0L)
+                .userIdleTimeoutSecs(10L)
                 .voicemailDetection(
                     TelephonySettings.VoicemailDetection.builder()
                         .onVoicemailDetected(
@@ -79,7 +80,8 @@ internal class TelephonySettingsTest {
             )
         assertThat(telephonySettings.supportsUnauthenticatedWebCalls()).contains(true)
         assertThat(telephonySettings.timeLimitSecs()).contains(30L)
-        assertThat(telephonySettings.userIdleTimeoutSecs()).contains(30L)
+        assertThat(telephonySettings.userIdleReplySecs()).contains(0L)
+        assertThat(telephonySettings.userIdleTimeoutSecs()).contains(10L)
         assertThat(telephonySettings.voicemailDetection())
             .contains(
                 TelephonySettings.VoicemailDetection.builder()
@@ -131,7 +133,8 @@ internal class TelephonySettingsTest {
                 )
                 .supportsUnauthenticatedWebCalls(true)
                 .timeLimitSecs(30L)
-                .userIdleTimeoutSecs(30L)
+                .userIdleReplySecs(0L)
+                .userIdleTimeoutSecs(10L)
                 .voicemailDetection(
                     TelephonySettings.VoicemailDetection.builder()
                         .onVoicemailDetected(