+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/versions/UpdateAssistantTest.kt
@@ -79,7 +79,8 @@ internal class UpdateAssistantTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(
@@ -298,7 +299,8 @@ internal class UpdateAssistantTest {
                     )
                     .supportsUnauthenticatedWebCalls(true)
                     .timeLimitSecs(30L)
-                    .userIdleTimeoutSecs(30L)
+                    .userIdleReplySecs(0L)
+                    .userIdleTimeoutSecs(10L)
                     .voicemailDetection(
                         TelephonySettings.VoicemailDetection.builder()
                             .onVoicemailDetected(
@@ -544,7 +546,8 @@ internal class UpdateAssistantTest {
                         )
                         .supportsUnauthenticatedWebCalls(true)
                         .timeLimitSecs(30L)
-                        .userIdleTimeoutSecs(30L)
+                        .userIdleReplySecs(0L)
+                        .userIdleTimeoutSecs(10L)
                         .voicemailDetection(
                             TelephonySettings.VoicemailDetection.builder()
                                 .onVoicemailDetected(