+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/models/ai/assistants/versions/VersionUpdateParamsTest.kt
@@ -78,7 +78,8 @@ internal class VersionUpdateParamsTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(
@@ -330,7 +331,8 @@ internal class VersionUpdateParamsTest {
                                 )
                                 .supportsUnauthenticatedWebCalls(true)
                                 .timeLimitSecs(30L)
-                                .userIdleTimeoutSecs(30L)
+                                .userIdleReplySecs(0L)
+                                .userIdleTimeoutSecs(10L)
                                 .voicemailDetection(
                                     TelephonySettings.VoicemailDetection.builder()
                                         .onVoicemailDetected(
@@ -572,7 +574,8 @@ internal class VersionUpdateParamsTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(