+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/services/async/ai/AssistantServiceAsyncTest.kt
@@ -88,7 +88,8 @@ internal class AssistantServiceAsyncTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(
@@ -346,7 +347,8 @@ internal class AssistantServiceAsyncTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(