+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/services/async/ai/assistants/VersionServiceAsyncTest.kt
@@ -114,7 +114,8 @@ internal class VersionServiceAsyncTest {
                                     )
                                     .supportsUnauthenticatedWebCalls(true)
                                     .timeLimitSecs(30L)
-                                    .userIdleTimeoutSecs(30L)
+                                    .userIdleReplySecs(0L)
+                                    .userIdleTimeoutSecs(10L)
                                     .voicemailDetection(
                                         TelephonySettings.VoicemailDetection.builder()
                                             .onVoicemailDetected(