+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/services/blocking/ai/AssistantServiceTest.kt
@@ -88,7 +88,8 @@ internal class AssistantServiceTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(
@@ -344,7 +345,8 @@ internal class AssistantServiceTest {
                             )
                             .supportsUnauthenticatedWebCalls(true)
                             .timeLimitSecs(30L)
-                            .userIdleTimeoutSecs(30L)
+                            .userIdleReplySecs(0L)
+                            .userIdleTimeoutSecs(10L)
                             .voicemailDetection(
                                 TelephonySettings.VoicemailDetection.builder()
                                     .onVoicemailDetected(