+++ telnyx-core/src/test/kotlin/com/telnyx/sdk/services/blocking/ai/assistants/VersionServiceTest.kt
@@ -113,7 +113,8 @@ internal class VersionServiceTest {
                                     )
                                     .supportsUnauthenticatedWebCalls(true)
                                     .timeLimitSecs(30L)
-                                    .userIdleTimeoutSecs(30L)
+                                    .userIdleReplySecs(0L)
+                                    .userIdleTimeoutSecs(10L)
                                     .voicemailDetection(
                                         TelephonySettings.VoicemailDetection.builder()
                                             .onVoicemailDetected(